Use global transform instead of collider transform for intended direction

+++ examples/other/src/implementations.hpp
@@ -152,8 +152,8 @@ struct MapInfo {
     glm::vec2 unit_pixel_dimensions;
     glm::vec2 coord_dimensions;
     glm::vec2 coord_origin;
-    glm::vec2 toCoords(glm::vec2 v);
-    glm::vec2 toPixels(glm::vec2 v);
+    glm::ivec2 toCoords(glm::vec2 v);
+    glm::vec2 toPixels(glm::ivec2 v);
     bool isValid(glm::vec2 v);
 };
